refactor(media): drop worker.ts side-channel; collection-resident _meta tools
Builds on the SDK's new `_meta` forwarding
(umbraco/Umbraco-MCP-Base#127) to consolidate the three previous tools:

- `create-media` now streams its URL path (was a separate
  `create-media-from-url-stream` registered out-of-band on McpServer).
  Multi-MB Drive/Dropbox/Unsplash uploads no longer trip the MCP
  transport timeout.
- `create-media-from-file` is back in the regular `media` collection
  with `_meta: { "openai/fileParams": ["file"] }` — once the SDK
  forwards `_meta` to `tools/list`, no `worker.ts` side-channel
  needed. (Note: tested in ChatGPT today; the connector reads the
  declaration and shows "Files X" in the consent dialog, but the
  automatic file-object injection isn't firing for ChatGPT-generated
  images in dev mode right now — file path remains useful for
  attached-by-user files. URL path is the recommended flow.)
- `worker.ts` shrinks to a single `setStreamingAuthContext` call
  (still needed because the streaming POST bypasses the orval
  transport to use `duplex: "half"`).

Streaming logic lives in
`src/umb-management-api/tools/media/post/helpers/streaming-upload.ts`
— Drive response.body pipes straight into a custom multipart
ReadableStream via fetch, with `duplex: "half"`. CPU stays near
zero; download + upload run concurrently.

Verified end-to-end against `umbraco-cms-dev-mcp-staging` via
ChatGPT:
- create-media sourceType=url, 188 KB Drive JPEG → success
- create-media sourceType=url, 3.57 MB Drive JPEG → success (the
  file that previously timed out via the buffered orval path)

Replaces the earlier worker.ts-registered helper tools — they're
deleted in this commit. Pre-existing auth-policies defensiveness
(?? []) carries over from the previous commit on this branch.

NOTE: package.json still pins beta.25 of @umbraco-cms/mcp-server-sdk
and @umbraco-cms/mcp-hosted; the `_meta`-forwarding fix needs a beta
publish from umbraco/Umbraco-MCP-Base#127 before this PR can land
against real npm. Locally tested by `npm install`-ing the packed
beta.25 .tgz from the patched base worktree.
