+++ src/umb-management-api/tools/media/index.ts
@@ -1,4 +1,5 @@
 import { createCreateMediaTool } from "./post/create-media.js";
+import CreateMediaFromFileTool from "./post/create-media-from-file.js";
 import { createCreateMediaMultipleTool } from "./post/create-media-multiple.js";
 import CreateMediaFolderTool from "./post/create-media-folder.js";
 import DeleteMediaTool from "./delete/delete-media.js";
@@ -63,6 +64,7 @@ export const MediaCollection: ToolCollectionExport = {
     if (AuthorizationPolicies.SectionAccessMedia(user)) {
       const allowFilePath = isFilePathUploadAllowed();
       tools.push(createCreateMediaTool({ allowFilePath }));
+      tools.push(CreateMediaFromFileTool);
       tools.push(createCreateMediaMultipleTool({ allowFilePath }));
       tools.push(CreateMediaFolderTool);
       tools.push(DeleteMediaTool);
