+++ src/worker.ts
@@ -25,11 +25,7 @@ import { umbracoCloudSiteRouting } from "@umbraco-cms/mcp-hosted/cloud";
 // CMS collections and registries
 import { collections, allModes, allModeNames, allSliceNames } from "./collections.js";
 import { UmbracoManagementClient } from "./umb-management-api/umbraco-management-client.js";
-import { registerCreateMediaFromFileTool } from "./umb-management-api/tools/media/post/create-media-from-file.js";
-import {
-  registerCreateMediaFromUrlStreamTool,
-  setStreamingAuthContext,
-} from "./umb-management-api/tools/media/post/create-media-from-url-stream.js";
+import { setStreamingAuthContext } from "./umb-management-api/tools/media/post/helpers/streaming-upload.js";
 
 // ============================================================================
 // Server Configuration
@@ -63,14 +59,12 @@ export class UmbracoMcpAgent extends McpAgent<HostedMcpEnv, unknown, AuthProps> 
       this.env,
       this.props!,
     );
-    // These two tools are registered directly on the McpServer (rather than
-    // through a regular collection) because they need extras the SDK's
-    // collection-registration loop doesn't currently forward — `_meta` for
-    // openai/fileParams in one case, the per-tenant token + base URL for the
-    // streaming POST in the other. See each module's header comment.
-    registerCreateMediaFromFileTool(this.server);
+    // create-media's streaming `url`/`file` source types bypass the orval
+    // transport (so they can use `duplex: "half"`). They need direct KV
+    // access for the OAuth token + the resolved per-tenant base URL — the
+    // tool handler doesn't receive env/props, so we stash them here. Safe
+    // because each Durable Object instance processes one request at a time.
     setStreamingAuthContext({ env: this.env, tokenKey: this.props!.umbracoTokenKey });
-    registerCreateMediaFromUrlStreamTool(this.server);
   }
 
   // Diagnostic: surface stack traces for the otherwise-opaque
